+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Fixed including and excluding keywords in paths.
 - More consistent information about frozen core electrons. Also changed
   `'frozencore'` to ``'frozen_core'``.
+- Parsing errors will not stop qcjson-creator.py if debug is False.
 
 ## [0.2.0] 2021-02-09